+++ backend/canisters/group/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Avoid retrying c2c call if recipient canister is uninstalled ([#7302](https://github.com/open-chat-labs/open-chat/pull/7302))
+- Fix `c2c_can_issue_access_token` ([#7366](https://github.com/open-chat-labs/open-chat/pull/7366))
 
 ## [[2.0.1581](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1581-group)] - 2025-01-24